Refactoring and improvment of backward compatibility tests.
- Everything is moved to quickwit-metastore as opposed to being in an independant crate.
- The code is factorized for the three different structure we check today.
- The logic that was within the build.rs file is now integrated into the unit tests. It is a bit clunky as unit tests are modifying our test data files, but it has the merit of being simpler and making it possible to spot the lack of update in the CI
- The expected.json files get updated automatically
- The lagging expected.json files are updated.
- demux_enabled is skipped upon serialization by default.
in test mode, the default create/update timestamps are constants.
- the backward compatibility unit tests are now regular unit tests as opposed to integration tests. (Integrations tests makes code re-use clunky)
