--- quickwit-backward-compat/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "quickwit-backward-compat"
-version = "0.1.0"
-edition = "2021"
-license = 'AGPL-3.0-or-later'  # For a commercial, license, contact hello@quickwit.io
-
-[dev-dependencies]
-anyhow = "1"
-once_cell = "1"
-quickwit-metastore = { path = "../quickwit-metastore" }
-serde_json = "1"
-
-[build-dependencies]
-anyhow = "1"
-byte-unit = { version = "4", default-features = false, features = ["serde"] }
-md5 = "0.7"
-quickwit-config = { path = "../quickwit-config" }
-quickwit-index-config = { path = "../quickwit-index-config" }
-quickwit-metastore = { path = "../quickwit-metastore" }
-serde_json = "1"
-toml = "0.5"
