--- quickwit-backward-compat/test-data/split-metadata/v0-a56cdae1f7333d023f1d6303ed5fd301.json
@@ -1,24 +1,0 @@
-{
-  "footer_offsets": {
-    "end": 2000,
-    "start": 1000
-  },
-  "split_metadata": {
-    "create_timestamp": 3,
-    "demux_num_ops": 1,
-    "num_docs": 12303,
-    "size_in_bytes": 234234,
-    "split_id": "split",
-    "split_state": "Published",
-    "tags": [
-      "234",
-      "aaa"
-    ],
-    "time_range": {
-      "end": 130198,
-      "start": 121000
-    },
-    "update_timestamp": 1
-  },
-  "version": "0"
-}
