+++ quickwit-metastore/test-data/file-backed-index/v0-cd9b0bf64dbb7af9e92bfed8a569c890.json
