+++ quickwit-metastore/test-data/index-metadata/unversioned.json
