+++ quickwit-metastore/test-data/index-metadata/v0-082493da53735bc58ec35c1648a4f14c.json
