+++ quickwit-metastore/test-data/index-metadata/v0-4a463f5078aa6acf314c95936336fa68.json
