+++ quickwit-metastore/test-data/index-metadata/v0-cb4b352caf8990d0e7a54582a4478cd5.json
