+++ quickwit-metastore/test-data/split-metadata/v0-219594cc7cb3d5b576e2d21346bf4979.expected.json
