+++ quickwit-metastore/test-data/split-metadata/v0-a56cdae1f7333d023f1d6303ed5fd301.expected.json
