+++ quickwit-metastore/test-data/split-metadata/v1-b9a036ccc850646c89bc008c65c4249f.json
@@ -1,0 +1,20 @@
+{
+  "create_timestamp": 3,
+  "demux_num_ops": 1,
+  "footer_offsets": {
+    "end": 2000,
+    "start": 1000
+  },
+  "num_docs": 12303,
+  "size_in_bytes": 234234,
+  "split_id": "split",
+  "tags": [
+    "234",
+    "aaa"
+  ],
+  "time_range": {
+    "end": 130198,
+    "start": 121000
+  },
+  "version": "1"
+}
